requirements.txt
torchvision==0.20.1
torch==2.5.1
typer==0.15.1
fastapi==0.115.6
uvicorn==0.34.0
scikit-learn==1.6.0
matplotlib==3.7.3
hydra-core==1.3.2
torch-tb-profiler==0.4.3
numpy==1.25.0
loguru==0.7.3
wandb==0.19.2
pytorch-lightning==2.5.0
pytest==8.3.4
coverage==7.6.10
pydantic==2.10.5
httpx==0.28.1
locust==2.32.6
onnx==1.17.0
onnxruntime==1.20.1
onnxscript==0.1.0
bentoml==1.3.19
streamlit==1.41.1
google_cloud_run==0.10.14
evidently==0.5.1
prometheus-client==0.21.1